Rename AdaptVQE files
